scripts: introduce configurable number of load gen threads
This change adds support for a --threads flag to wrk2/render.sh, and
adds the number of benchmark load generator threads to benchmark.sh
and benchmark-multi.sh.

The change furthermore introduces sane defualts for RPS, app count,
and nr of threads when run on a minimal cluster of 2 t1.small workers.

Signed-off-by: Thilo Fromm <thilo@kinvolk.io>